Update Program.cs
Notes: 
With lines 72 - 78 commented out, I am able to save a copy of the file to a new location using the File.Copy method.

However, I am getting into a muddle when asking the user to input the file. If you comment these back in and use the string targetPathUser in place of targetPath; various exceptions are thrown to indicate I either don't have permission, the file does not exist or that it reads C:\\users\\file.

I am also unsure whether line 78 'newPlace = "@" + targetPathUser;' is the correct way to attempt to append the @ to the path the user inputs. I have done this on the assumption that the user will not be aware of escape characters in order to type \\.

The 'done' user input ideally would not be added to the list of to-dos. I am unsure how to ignore this, perhaps do something with an array, and remove the last item....